feat: CafeInfoModal에서 CafeSummary를 분리
