--- client/src/components/CafeInfoModal.stories.tsx
@@ -1,19 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import CafeInfoModal from './CafeInfoModal';
-
-type Story = StoryObj<typeof CafeInfoModal>;
-
-const meta: Meta<typeof CafeInfoModal> = {
-  title: 'CafeInfoModal',
-  component: CafeInfoModal,
-};
-
-export default meta;
-
-export const Default: Story = {
-  args: {
-    title: '성수동 카페',
-    address: '서울 성동구 연무장3길 6',
-    content: '안녕하세요~!!!!!',
-  },
-};
